pyproject.toml
[project]
name = "sway-out"
version = "0.0.0"
description = "A layout save/apply utility for the Sway window manager."
authors = []
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
"click (>=8.2.1,<9.0.0)",
"pydantic (>=2.11.7,<3.0.0)",
"pyyaml (>=6.0.2,<7.0.0)",
"i3ipc (>=2.2.1,<3.0.0)"
]


[project.scripts]
sway-out = "sway_out.main:main"


[tool.poetry]
packages = [{ include = "sway_out", from = "src" }]


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
commitizen = "^4.8.3"
pre-commit = "^4.2.0"


[tool.poetry.group.docs.dependencies]
mkdocs = "^1.6.1"
mkdocstrings = "^0.29.1"
mkdocs-material = "^9.6.14"
mkdocs-autorefs = "^1.4.2"


[tool.poetry.group.test.dependencies]
pytest = "^8.4.0"
pytest-asyncio = "^1.0.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
log_cli_level = "DEBUG"


[tool.black]


[tool.isort]
profile = "black"
src_paths = ["src", "tests"]
skip_gitignore = true


[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$major.$minor.$patch$prerelease"
version_scheme = "pep440"
version_provider = "poetry"
update_changelog_on_bump = true
major_version_zero = true
#changelog_start_rev = "v0.1.0"
changelog_incremental = false
allowed_prefixes = ["fixup!", "squash!"]
